requirements.txt
pywebview>=4.0
pyaudio>=0.2.13
keyboard>=0.13.5
pyperclip>=1.8.2
requests>=2.28.0
